Refactor create and update methods
Since the validate() method already returns an array of the existing form
attributes, we can use this in
the Article::create method
Passing the validate method into it is the same but without repeating the
same code.

Furthuremore, since we use the same validation for multiple methods we
can pull it into its own method and reuse it.

Laravel guards against vulnerabilities by requiring you to specify
which fields can be updated. In this case we set $guarded to emply array which
allows all items (not secure)