EP28 Named Routes
In this lesson I learned how to create named routes.
This avoids the pitfal of having to change all the routes in a project
when the path needs to change, only need to update once in the routes
file.
Additionally, we created a custom path() method in our Articles Model
which makes the code just a little less verbos.