EP29-30 Foreign Keys and DB Factories
In these modules I set up an Article some Eloquent relations in the
Aarticles model.
Created a foreign key user_id in the articles  table so that
orphaned articles are deleted when a user is deleted.
Additionally created an Article Factory which may be used to quickly
generate random articles in php Tinker using the Faker library.