EP33 Attach and validate many-many relationships
Used the belongsToMany relationship created in previous episode to add
tag multi-selectbox to the articles/create form.
When returning the view, we pass in tags as a variable.
The crete form now has a simple multi-selectbox populated with all the
tags.

Validation:
We do not want to allow users to get to the point of SQL error, thus we
need to ad a new "exists" validation in the validateArticles() method.
Because this now references a different table, the store() method must
be refactored to use the new attach() method.