EP49 Events
Sometimes we want to seperate out the core logic of the Controller and
its sideeffects. Eg after a user purchases a product we want to
send an email confirmations, award them an achivement, and queue up an
email to be sent with a discount code in the future.

While this can be done procedurally in the Controller this may end up
being very long. There is where Events come into play.

Created a ProductPurchased event using:
`php artisan make:event ProductPurchased`
Some uncessary code was stripped from this class.
Now we can dispatch this event in 2 different ways, most commonly with
the 'event()' method.
Next we created several listeners:
`php artisan make:listener AwardAchievement -e ProductPurchased`
The `-e` flag simply points our listener to the ProductPurchased class.
Lastly we needed to hook each listener with with the new event.
This can either be done verbose in the the EventServiceProvider.php
listen array. Or we can have laravel do some of its magic and override
the 'shouldDiscoverEvents()' method to return true.
Now each listener will be paired with the corresponding event.