EP50 Authorization
In this section we created  some authorization structure around the
conversation replies "mark as best" functionality.

On the view, we used the blade '@can' durective which used the
AuthServiceProvider.php class.
Additionally, created a new controller since the functionality is not core to
Conversation itself:
`php artisan make:controller ConversationBestReplyController`
Here we created a store()
method which used  Laravel's authorize()
We then defined a Gate::define() which returns true if the user of the
conversation matches the current logged in user.

This is an OK approach for small apps, however using Policies is often
more favorable.
Thus we removed the Gate::define() and instead scaffolded a Policy:
`php artisan make:policy ConversationPolicy --model=Conversation`
This comes preloaded with several methods, but we only need 'update()'
where we place the same logic as we had in the Gate definition.

Lastly we pulled out some common functionality from the controller and
into the model - setBestReply() - as well as isBest() from the view.